[BUG] whoseTurnNo반영 오류 수정 (#115)
