+++ src/main/java/com/memorytrace/controller/DiaryController.java
@@ -72,16 +72,4 @@ public class DiaryController {
         return new ResponseEntity<>(
             DefaultRes.res(StatusCode.OK, ResponseMessage.READ_DIARY_DETAIL, diary), HttpStatus.OK);
     }
-
-
-    @ApiOperation(value = "해당 유저 다이어리 나가기")
-    @ApiResponses(value = {
-        @ApiResponse(code = 200, message = "해당 유저 다이어리 나가기 성공")
-    })
-    @PutMapping("exit/{bid}")
-    public ResponseEntity<DefaultRes<DiaryDetailResponseDto>> exitDiary(@PathVariable Long bid) {
-        diaryService.exitDiary(bid);
-        return new ResponseEntity<>(
-            DefaultRes.res(StatusCode.OK, ResponseMessage.EXIT_DIARY), HttpStatus.OK);
-    }
 }
